# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A meow API server built with Go, inspired by meow",
        "title": "meow meow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "meow API Support",
            "url": "https://github.com/your-username/meow",
            "email": "support@meow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/session/{sessionId}/privacy/blocklist": {
            "get": {
                "description": "Get the list of blocked contacts (blocklist)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privacy"
                ],
                "summary": "Get blocked contacts list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Block or unblock a contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privacy"
                ],
                "summary": "Update blocklist (block/unblock contact)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update blocklist request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBlocklistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.BlocklistResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/privacy/find": {
            "post": {
                "description": "Get specific privacy settings or all settings if none specified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privacy"
                ],
                "summary": "Find specific privacy settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Find privacy settings request (optional)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.FindPrivacySettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/privacy/set": {
            "put": {
                "description": "Set multiple privacy settings in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privacy"
                ],
                "summary": "Set multiple privacy settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Privacy settings request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetAllPrivacySettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivacySettingsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BlocklistResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error": {
                    "$ref": "#/definitions/dto.PrivacyErrorResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.FindPrivacySettingsRequest": {
            "type": "object",
            "properties": {
                "settings": {
                    "description": "Specific settings to retrieve: [\"groupAdd\", \"lastSeen\", \"status\", \"profile\", \"readReceipts\", \"callAdd\", \"online\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.PrivacyErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.PrivacySettingsData": {
            "type": "object",
            "properties": {
                "callAdd": {
                    "description": "Who can call: \"all\", \"known\"",
                    "type": "string"
                },
                "groupAdd": {
                    "description": "Who can add to groups: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "lastSeen": {
                    "description": "Who can see last seen: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "online": {
                    "description": "Who can see online status: \"all\", \"match_last_seen\"",
                    "type": "string"
                },
                "profile": {
                    "description": "Who can see profile photo: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "readReceipts": {
                    "description": "Read receipts: \"all\", \"none\"",
                    "type": "string"
                },
                "status": {
                    "description": "Who can see status: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                }
            }
        },
        "dto.PrivacySettingsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.PrivacySettingsData"
                },
                "error": {
                    "$ref": "#/definitions/dto.PrivacyErrorResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.SetAllPrivacySettingsRequest": {
            "type": "object",
            "properties": {
                "callAdd": {
                    "description": "Who can call: \"all\", \"known\"",
                    "type": "string"
                },
                "groupAdd": {
                    "description": "Who can add to groups: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "lastSeen": {
                    "description": "Who can see last seen: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "online": {
                    "description": "Who can see online status: \"all\", \"match_last_seen\"",
                    "type": "string"
                },
                "profile": {
                    "description": "Who can see profile photo: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                },
                "readReceipts": {
                    "description": "Read receipts enabled/disabled",
                    "type": "boolean"
                },
                "status": {
                    "description": "Who can see status: \"all\", \"contacts\", \"contact_blacklist\", \"none\"",
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateBlocklistRequest": {
            "type": "object",
            "required": [
                "action",
                "jid"
            ],
            "properties": {
                "action": {
                    "description": "\"block\" or \"unblock\"",
                    "type": "string"
                },
                "jid": {
                    "description": "JID to block/unblock",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Simply provide your API key directly: \"YOUR_API_KEY\". The system automatically detects if it's a Global API Key (can access all sessions and session management) or a Session-specific API Key (can only access the specific session it belongs to).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}